pyproject.toml
[tool.poetry]
name = "keyboard-warrior"
version = "0.1.0"
description = ""
authors = ["stanluuk31 <467804@student.saxion.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
blessed = "^1.20.0"
pyfiglet = "^1.0.2"
pypi = "^2.1"
pydub = "^0.25.1"
python-vlc = "^3.0.20123"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
